fix: unconditional class level reset in spell pool and adventure.cpp narrowing cast
- generate_spell_pool.pl: unconditionally set all class fields to 1 so
  values of 0 and 255 (both meaning 'cannot use') are also cleared,
  fixing 'not high enough level to memorize' for all spells
- export_spells_us.pl: chmod only (no logic change)
- code/world/adventure.cpp: static_cast<uint32_t> on instance_id to
  fix narrowing conversion build error
- server/shared/spells: regenerated spell pool binary

+++ code/world/adventure.cpp
@@ -71,7 +71,7 @@ void Adventure::AddPlayer(std::string character_name, bool add_client_to_instanc
 			AdventureMembersRepository::ReplaceOne(
 				database,
 				AdventureMembersRepository::AdventureMembers{
-					.id = instance_id,
+					.id = static_cast<uint32_t>(instance_id),
 					.charid = static_cast<uint32_t>(character_id)
 				}
 			);
